feat: check transformation result when typing
This means that the transformation function for the current argument
will be executed on the current text for the argument. If an error
occurs during transformation, it will be displayed in the argument
suggestion.

There is also now an `expensive` flag for types, indicating if this
functionality should be disabled for the type.