feat: configurable command class construction
This allows the way in which command classes are constructed to be configured.

This is especially useful when using Flamework because it means you can pass dependencies to command classes by using `Modding.createDependency` to construct the class. The main caveat here is that you need a custom `@Register` decorator with the `flamework:parameters` JSDoc tag.