fix: remove config freezing
This was in place to prevent modification of the config object, since only some properties can be updated.

This causes issues when using defaultContextState however, when providing state that shouldn't be frozen (e.g. a Matter world).

The config can still be enforced as readonly at the type-level, though this is not currently possible to replicate with Luau types yet.